refactor: 정리 - login 관련 정리 - enum 위치 변경
