+++ src/main/java/com/doremi/marketdoremi/domain/authority/repository/AuthorityRepository.java
@@ -1,7 +1,7 @@
 package com.doremi.marketdoremi.domain.authority.repository;
 
 import com.doremi.marketdoremi.domain.authority.entity.Authority;
-import com.doremi.marketdoremi.web.Role;
+import com.doremi.marketdoremi.codes.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AuthorityRepository extends JpaRepository<Authority, Role> {
