+++ src/main/java/com/doremi/marketdoremi/domain/grade/entity/MemberGrade.java
@@ -1,13 +1,12 @@
 package com.doremi.marketdoremi.domain.grade.entity;
 
 import javax.persistence.Embedded;
-import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
 import com.doremi.marketdoremi.domain.benefit.entity.Benefit;
-import com.doremi.marketdoremi.web.Grade;
+import com.doremi.marketdoremi.codes.Grade;
 
 @Table(name = "grade")
 //@Entity
