+++ src/main/resources/static/rest/member.http
@@ -26,8 +26,8 @@ Content-Type: application/json
 
 {
   "member": {
-    "memberId": "admin1",
-    "password": "1234"
+    "memberId": "admintest",
+    "password": "admin1234"
   },
   "roles": ["ADMIN", "USER"],
   "memberInfo": {
@@ -48,26 +48,8 @@ Content-Type: application/json
 
 < ./signup.json
 
-### 로그인
-GET http://localhost:9000/login
-Content-Type: application/json
-
-{
-  "memberId": "user1231234",
-  "password": "123456"
-}
-
 ### 회원가입
 POST http://localhost:9000/api/v1/member
 Content-Type: application/json
 
 < ./signup.json
-
-### 로그인
-GET http://localhost:9000/login
-Content-Type: application/json
-
-{
-  "memberId": "user1231234",
-  "password": "123456"
-}
